+++ examples/range_zip_example.cpp
@@ -19,7 +19,7 @@ int main() {
 #ifndef NO_LIBZIP
   std::array const shape{z.size()};
   cnpypp::npz_save("range_zip_data.npz", "struct", {"a", "b", "c"}, z.begin(),
-                   cnpypp::span<size_t const>{shape.data(), shape.size()});
+                   cnpypp::span<uint64_t const>{shape.data(), shape.size()});
 #endif
 
   return EXIT_SUCCESS;
