+++ src/c_interface.c
@@ -8,7 +8,7 @@
 int cnpypp_npy_save_1d(char const* fname, enum cnpypp_data_type dtype,
                        void const* start, size_t const num_elem,
                        char const* mode) {
-  size_t const shape[] = {num_elem};
+  uint64_t const shape[] = {num_elem};
   return cnpypp_npy_save(fname, dtype, start, shape, 1, mode,
                          cnpypp_memory_order_c);
 }
@@ -17,7 +17,7 @@ int cnpypp_npy_save_1d(char const* fname, enum cnpypp_data_type dtype,
 int cnpypp_npz_save_1d(char const* zipname, char const* fname,
                        enum cnpypp_data_type dtype, void const* data,
                        size_t num_elem, char const* mode) {
-  size_t const shape[] = {num_elem};
+  uint64_t const shape[] = {num_elem};
   return cnpypp_npz_save(zipname, fname, dtype, data, shape, 1, mode,
                          cnpypp_memory_order_c);
 }
